Add profile support

+++ ProjectEarthServerAPI/Models/Features/CraftingResponse.cs
@@ -48,7 +48,7 @@ namespace ProjectEarthServerAPI.Models.Features
         public string state { get; set; } // State: Active, Completed or Locked
         public BoostState boostState { get; set; } // See class below
         public UnlockPrice unlockPrice { get; set; } // Price to unlock item
-        public int streamVersion { get; set; } // StreamVersion with changes, unused in our version
+        public long streamVersion { get; set; } // StreamVersion with changes, unused in our version
     }
 
     public class BoostState
